Reorganize files into better structure.
