+++ src/core/simulation.js
