+++ src/core/stats.js
