+++ src/ui/fibonacci-config.js
